Add XDSL Compiler Flow (kuleuven-micas/snax-mlir#9)
This pull request adds the XDSL project as a submodule to write custom MLIR passes for lowering linalg->snax.
It also includes a first simple pass that detects elementwise integer multiplications, and lowers these to an external function call. This external function call can then be an implementation using a snax accelerator.